requirements.txt
yfinance
python-binance
schedule
ta
xgboost==1.4.0
tensorflow
keras